Balancing Project
Added Tooltips for Firework Block and Mese Box and Radsensor
Added Particle Accelerator to structure Wand
Added Keep-inventory to Crates
Added Keep-power to Batteries
Added Wire-Oredict
Added Euphemium Cladding
Added Biomass Chemplant recipe
Added Tooltip to DFC cores
Added UNDEFINED DFC lens
Added custom RBMK Fuel rod channel colors and cherenkov
Added Schrabidium and Yharonite SC Batteries
Added new title image

Changed the name from Alcater Edition to Extended Edition
Changed a some textures of items blocks and models
Changed Fallout to hopefully not crash when replacing leaves
Changed Void Bullet to not cause an explosion
Changed both centrifuges to use the new models
Changed Centrifuge gui to the new one
Changed moved verticium and above to Hadron
Changed all DFC recipes
Changed DFC crate workings - now it needs some time to process and more SPK makes it quicker
Changed DFC Catalyst to be more balanced
Changed the Hadron to work exactly as in 1.7.10
Changed some Hadron recipes to 1.7.10
Changed Electronium book recipe to 1.7.10
Changed Silex recipes to 1.7.10 and cleaned the rest
Changed a lot of irradiation channel recipes to become more balanced
Changed Marshmallows to be edible and have a tooltip
Changed DFC Lens power draw and lens stats
Changed Ivy Mike model to be 1.7.10
Changed DFC explosion to be jammable
Changed DFC emitter to break lower than 10k resistance rather then 6k
Changed DFC Catalysts and cores to have DFC in the name
Changed a lot of fluid textures to be 1.7.10

Fixed alot of assembler recipes and updated others to 1.7.10
Fixed Exo warheads not removing snow
Fixed Moderated Fuel rod reflection flux type
Fixed Items having Coal or Asbestos effect that should not
Fixed Fel laser being behind other models
Fixed Steam and Water info in RBMK rods that have neither
Fixed ArcFurnace slot management

Removed Dynospheres
Removed unused textures

+++ src/main/java/com/hbm/tileentity/machine/TileEntityAMSBase.java
@@ -251,9 +251,9 @@ public class TileEntityAMSBase extends TileEntity implements ITickable, ISource,
 						fuelMod *= ItemCatalyst.getFuelMod(inventory.getStackInSlot(i));
 					}
 
-					powerBase = ItemAMSCore.getPowerBase(inventory.getStackInSlot(12));
-					heatBase = ItemAMSCore.getHeatBase(inventory.getStackInSlot(12));
-					fuelBase = ItemAMSCore.getFuelBase(inventory.getStackInSlot(12));
+					powerBase = (int)ItemAMSCore.getPowerBase(inventory.getStackInSlot(12));
+					heatBase = (int)ItemAMSCore.getHeatBase(inventory.getStackInSlot(12));
+					fuelBase = (int)ItemAMSCore.getFuelBase(inventory.getStackInSlot(12));
 					
 					powerBase *= this.efficiency;
 					powerBase *= Math.pow(1.25F, booster);
